change structure, rename namespace, update some observer & factories
-change namespace
- group up some class & interface
- change observers into events including observers, subjects, registers
- use registers in service

+++ CafeManagement/CafeManagement/Models/Stock/StockUsageDetail.cs
@@ -5,7 +5,7 @@
         public Guid Id { get; set; }
         public Guid StockEntryDetailId { get; set; }  
         public Guid StockUsageLogId {  get; set; }
-        public double QuantityUsed { get; set; }
+        public float QuantityUsed { get; set; }
         public decimal TotalValue { get; set; } 
         public Guid UsageLogId { get; set; }
         public StockEntryDetail StockEntryDetail { get; set; }
